post update refactoring

+++ server/controllers/categories/actionsCategory.js
@@ -1,38 +1,73 @@
 const db = require('../../models/index');
+const validationError = require('../../services/validationError');
 
 const categoryTable = db.category;
 const languageTable = db.languages;
 module.exports = {
 	addCategory: async (body) => {
 		try {
-			return await categoryTable.create({ ...body });
+			const result = await categoryTable.create({ ...body });
+			return {
+				statusCode: 200,
+				message: result,
+			};
 		} catch (e) {
 			console.log(e);
-			return e;
+			return validationError(e);
 		}
 	},
 	deleteCategory: async (categoryId) => {
 		try {
-			await categoryTable.destroy({
+			const resultFind = await categoryTable.findOne({
 				where: {
 					id: categoryId,
 				},
 			});
-			return { message: 'Success' };
+			if (resultFind !== null) {
+				await categoryTable.destroy({
+					where: {
+						id: categoryId,
+					},
+				});
+				return {
+					statusCode: 200,
+					message: 'Category whis this id removed successfuly',
+				};
+			}
+
+			return {
+				message: 'Category whis this id not exists',
+				statusCode: 401,
+			};
 		} catch (e) {
 			console.log(e);
-			return e;
+			return validationError(e);
 		}
 	},
 	editCategory: async (categoryId, body) => {
 		try {
-			return await categoryTable.update(
-				{ ...body },
-				{ where: { id: categoryId } }
-			);
+			const resultFind = await categoryTable.findOne({
+				where: {
+					id: categoryId,
+				},
+			});
+			if (resultFind !== null) {
+				await categoryTable.update(
+					{ ...body },
+					{ where: { id: categoryId } }
+				);
+				return {
+					statusCode: 200,
+					message: 'Category whis this id updated successfuly',
+				};
+			}
+			return {
+				message: 'Category whis this id not exists',
+				statusCode: 401,
+			};
 		} catch (e) {
 			console.log(e);
-			return e;
+			return validationError(e);
 		}
 	},
 	getCategorys: async (language) => {
@@ -42,17 +77,27 @@ module.exports = {
 					iso_639_code: language,
 				},
 			});
-			return await categoryTable.findAll({
-				where: {
-					languageId: resultLang.id,
-				},
-				attributes: {
-					exclude: ['languageId'],
-				},
-			});
+			if (resultLang !== null) {
+				const result = await categoryTable.findAll({
+					where: {
+						languageId: resultLang.id,
+					},
+					attributes: {
+						exclude: ['languageId'],
+					},
+				});
+				return {
+					statusCode: 200,
+					message: result,
+				};
+			}
+			return {
+				statusCode: 401,
+				message: 'There is no category in this language',
+			};
 		} catch (e) {
 			console.log(e);
-			return e;
+			return validationError(e);
 		}
 	},
 };
