+++ server/controllers/submenus/actionsSubmenu.js
@@ -27,37 +27,72 @@ module.exports = {
 					{ ...body },
 					{ where: { id: submenuId } }
 				);
-				return;
+				return {
+					statusCode: 200,
+					message: `Submenu id ${submenuId} changed successfully`,
+				};
 			}
-
-			return;
+			return {
+				statusCode: 401,
+				message: 'Submenu with this id not exists',
+			};
 		} catch (error) {
 			console.log(error);
-			return error;
+			return validationError(error);
 		}
 	},
 	submenuBindParentMenu: async ({ submenuId, langId, parentMenuId }) => {
 		try {
-			return await menuSubmenuLangTable.create({
-				menuId: parentMenuId,
-				submenuId: submenuId,
-				languageId: langId,
-			});
-		} catch (error) {
-			console.log(error);
-			return error;
-		}
-	},
-	delSubmenu: async (submenuId) => {
-		try {
-			return await submenuTable.destroy({
+			const resultFind = await submenuTable.findOne({
 				where: {
 					id: submenuId,
 				},
 			});
+			if (resultFind !== null) {
+				await menuSubmenuLangTable.create({
+					menuId: parentMenuId,
+					submenuId: submenuId,
+					languageId: langId,
+				});
+				return {
+					statusCode: 200,
+					message: `Submenu whis this id change succesfully`,
+				};
+			}
+			return {
+				statusCode: 401,
+				message: 'Submenu whis this id not exists',
+			};
+		} catch (error) {
+			console.log(error);
+			return validationError(error);
+		}
+	},
+	delSubmenu: async (submenuId) => {
+		try {
+			const resultFind = await submenuTable.findOne({
+				where: {
+					id: submenuId,
+				},
+			});
+			if (resultFind !== null) {
+				await submenuTable.destroy({
+					where: {
+						id: submenuId,
+					},
+				});
+				return {
+					statusCode: 200,
+					message: 'Submenu whis this id removed succesfully',
+				};
+			}
+			return {
+				statusCode: 401,
+				message: 'Submenu whis this id not exists',
+			};
 		} catch (e) {
 			console.log(e);
-			return e;
+			return validationError(e);
 		}
 	},
 };
